Make `pew in venv` start a subshell
Currently `pew in myvenv` is an invalid command. It makes sense for
it to run a sensible default command in the virtual env, and a
sensible default command would be a subshell. This proxies that
currently invalid command to be the same as `pew workon myvenv`.

Make a straightforward duplication of test_workon to test the same
feature of pew in.