Make inve default command a shell
Which shell is platform-specific. If on windows, it will open a new
powershell. Otherwise, it is based on the 'SHELL' environment variable.

The logic for choosing a shell is still needed in the shell() function
for now because the shell_check uses it. However, it's not needed on
windows, because the check doesn't happen there.

If the shell to run is chosen automatically (the default), then write
the message notifying the user that a subshell is being entered.