Change the subshell entering message
The new message gives both the environment that we're entering, and
tells what command they should expect to see running. However, I would
ideally like to see all message in this normal case to be completely
removed.

I don't want any of that text to be part of the expected UI of the tool.
I do recognize that it can be helpful text for a beginner, but I think
that the noise in the general case makes it a net loss.

I also think that if the prompt can is modified appropriately to
indicate which environment the user is in, then it becomes easy for the
user to tell that they are in a virtual environment.