Beefing up tests so Tox works for 3.5+ on macOS
Need a newer Virtualenv for 3.6, and newer Pytest for 3.5+, so the
requirements.txt are modified to allow newer versions.

Path of tmpdir on macOS is a symlink and need to be resolved explicitly
to prevent the check from failing.

Some tests are beefed up to help pinpoint the problem more accurately.