Implement venv backend
The venv backend is used by default when "venv" and "ensurepip" are
available (i.e. Python 3.4+). This can be disabled with the
PEW_USE_VIRTUALENV environ.

Certain tests are skipped:

@skip_venv
Test makes assumptions that do not fit venv's structure. The
functionality is implemented, but the check is logically incompatible
with how venv does things.

@skip_venv_site_packages
Same as above, but specifically this has something to do with venv's
central configuration file, and can be fixed by implementing a parallel
test.

@skip_venv_cp
This uses the "pew cp" command, which is not implemented yet. I will try
to eliminate them if possible.